Can successfully move from marketing to home fragment while passing in User Information.
