+++ app/src/main/java/com/lambda/mnemecards/marketing/MarketingViewModel.kt
@@ -1,0 +1,10 @@
+package com.lambda.mnemecards.marketing
+
+import androidx.lifecycle.ViewModel
+import com.firebase.ui.auth.AuthUI
+
+class MarketingViewModel : ViewModel(){
+
+
+
+}
