--- app/src/main/java/com/lambda/mnemecards/marketing/marketingFragment.kt
@@ -1,26 +1,0 @@
-package com.lambda.mnemecards.marketing
-
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-import com.lambda.mnemecards.R
-
-/**
- * A simple [Fragment] subclass.
- */
-class marketingFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_marketing, container, false)
-    }
-
-
-}
