+++ app/src/main/java/com/lambda/mnemecards/overview/HomeFragment.kt
@@ -12,6 +12,7 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import androidx.navigation.fragment.findNavController
+import androidx.navigation.fragment.navArgs
 import com.facebook.CallbackManager
 import com.facebook.FacebookCallback
 import com.facebook.FacebookException
@@ -31,44 +32,22 @@ import com.google.firebase.firestore.SetOptions
 import com.lambda.mnemecards.R
 import com.lambda.mnemecards.databinding.FragmentHomeBinding
 import com.lambda.mnemecards.network.User
+import com.lambda.mnemecards.settings.SettingsViewModel
+import com.lambda.mnemecards.settings.SettingsViewModelFactory
 import java.util.*
 
-// For accessing Firestore
-var db = FirebaseFirestore.getInstance()
+
 
 class HomeFragment : Fragment() {
 
-    companion object {
-        const val TAG = "MainFragment"
-        const val SIGN_IN_RESULT_CODE = 1001
-    }
-
-    lateinit var fragmentContext: Context
-    var googleSignInClient: GoogleSignInClient? = null
-    val RC_SIGN_IN = 1000
-
-
-
     private var loggedInFlag: Boolean = false
-
-    // Name to be passed to the settings
-    var name: String? = "First Name"
-
-    // Photo URL to be passed to the settings to draw the picture.
-    var photoUrl: String? = "asd"
-
-    // User's Id
-    var userId: String?= "No Token"
-
-    var user: User = User()
-
-    private lateinit var auth: FirebaseAuth
 
     var callbackManager = CallbackManager.Factory.create()
 
-    private val viewModel: HomeViewModel by lazy {
-        ViewModelProviders.of(this).get(HomeViewModel::class.java)
-    }
+    private lateinit var viewModel: HomeViewModel
+    private lateinit var viewModelFactory: HomeViewModelFactory
+
+    private lateinit var fragmentContext: Context
 
     /**
      * Inflates the layout with Data Binding, sets its lifecycle owner to the OverviewFragment
@@ -78,6 +57,12 @@ class HomeFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+
+        val homeFragmentArgs by navArgs<HomeFragmentArgs>()
+
+        viewModelFactory = HomeViewModelFactory(homeFragmentArgs.name, homeFragmentArgs.photoUrl, homeFragmentArgs.user)
+        viewModel = ViewModelProviders.of(this, viewModelFactory)
+            .get(HomeViewModel::class.java)
 
         val binding = FragmentHomeBinding.inflate(inflater)
 
@@ -107,7 +92,7 @@ class HomeFragment : Fragment() {
         // Code that pops up the possible log in options
 
         if (!loggedInFlag) {
-            launchSignInFlow()
+//            launchSignInFlow()
             loggedInFlag = true
         }
 
@@ -129,29 +114,9 @@ class HomeFragment : Fragment() {
 //            .addOnSuccessListener { Log.d(TAG, "DocumentSnapshot successfully written!") }
 //            .addOnFailureListener { e -> Log.w(TAG, "Error writing document", e) }
 
-        // To get data
-        db.collection("Users").document("4uR0bkDdUeOvQdolIXbiP0kxLZs1")
-            .get()
-            .addOnCompleteListener { result ->
-                val document = result.result
-                Log.d("HomeFragment", "${document?.data}")
 
-            }
-            .addOnFailureListener { exception ->
-                Log.d("Get Deck", "Error getting documents: ", exception)
-            }
 
-            .addOnFailureListener { exception ->
-                Log.d("Get Deck", "Error getting documents: ", exception)
-            }
 
-        // To get data using custom objects
-        val docRef = db.collection("Users").document("4uR0bkDdUeOvQdolIXbiP0kxLZs1")
-        docRef.get().addOnSuccessListener { documentSnapshot ->
-            val user = documentSnapshot.toObject(User::class.java)
-            Log.i("HomeFragment2", user?.favSubjects)
-            Log.i("HomeFragment2", user?.mobileOrDesktop)
-        }
 
 
         setHasOptionsMenu(true)
@@ -191,24 +156,6 @@ class HomeFragment : Fragment() {
 //                Log.d("Get Deck", "Error getting documents: ", exception)
 //            }
 
-        //This code is to get decks
-        db.collection("DemoDeck").document("I2r2gejFYwCQfqafWlVy")
-            .get()
-            .addOnSuccessListener { result ->
-                Log.d(
-                    "Get Deck",
-                    "${result.id} => ${result.data} => ${result.reference} => ${result.metadata} =>"
-                )
-
-            }
-            .addOnFailureListener { exception ->
-                Log.d("Get Deck", "Error getting documents: ", exception)
-            }
-
-            .addOnFailureListener { exception ->
-                Log.d("Get Deck", "Error getting documents: ", exception)
-            }
-
 //        db.collection("DemoDeck").document().collection("Biology").get()
 //            .addOnSuccessListener { document ->
 //                if (document != null) {
@@ -220,14 +167,14 @@ class HomeFragment : Fragment() {
 //            .addOnFailureListener { exception ->
 //                Log.d("Get Deck", "get failed with ", exception)
 //            }
-        var gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-            .requestIdToken(getString(R.string.default_web_client_id))
-            .requestEmail()
-            .build()
-        googleSignInClient = GoogleSignIn.getClient(fragmentContext, gso)
+//        var gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+//            .requestIdToken(getString(R.string.default_web_client_id))
+//            .requestEmail()
+//            .build()
+//        googleSignInClient = GoogleSignIn.getClient(fragmentContext, gso)
 
         // Initialize Firebase Auth
-        auth = FirebaseAuth.getInstance()
+//        auth = FirebaseAuth.getInstance()
 
 //        btn_google_login.setOnClickListener {
 //            signIn()
@@ -281,129 +228,69 @@ class HomeFragment : Fragment() {
 //        }
     }
 
-    private fun signIn() {
+//    private fun signIn() {
+//
+//        val signInIntent = googleSignInClient?.signInIntent
+//        startActivityForResult(signInIntent, RC_SIGN_IN)
+//
+//    }
 
-        val signInIntent = googleSignInClient?.signInIntent
-        startActivityForResult(signInIntent, RC_SIGN_IN)
 
-    }
-
-
-    fun firebaseAuthWithGoogle(acct: GoogleSignInAccount) {
-        var credential = GoogleAuthProvider.getCredential(acct?.idToken, null)
-        FirebaseAuth.getInstance().signInWithCredential(credential).addOnCompleteListener { task ->
-            if (task.isSuccessful) {
-                println("Google Login Success")
-                val user = auth.currentUser
-                val message = " ${user?.displayName} + ${user?.email} + ${user?.photoUrl}"
-
-                Log.i("INFORMATION GOOGLE", message)
-            }
-        }
-    }
+//    fun firebaseAuthWithGoogle(acct: GoogleSignInAccount) {
+//        var credential = GoogleAuthProvider.getCredential(acct?.idToken, null)
+//        FirebaseAuth.getInstance().signInWithCredential(credential).addOnCompleteListener { task ->
+//            if (task.isSuccessful) {
+//                println("Google Login Success")
+//                val user = auth.currentUser
+//                val message = " ${user?.displayName} + ${user?.email} + ${user?.photoUrl}"
+//
+//                Log.i("INFORMATION GOOGLE", message)
+//            }
+//        }
+//    }
 
     override fun onStart() {
         super.onStart()
 
-        val currentUser = auth.currentUser
+//        val currentUser = auth.currentUser
     }
 
-    fun facebookLogin() {
-        LoginManager.getInstance()
-            .logInWithReadPermissions(this, Arrays.asList("public_profile", "email"))
-        LoginManager.getInstance()
-            .registerCallback(callbackManager, object : FacebookCallback<LoginResult> {
-                override fun onSuccess(result: LoginResult?) {
-                    firebaseAuthWithFacebook(result)
-                }
+//    fun facebookLogin() {
+//        LoginManager.getInstance()
+//            .logInWithReadPermissions(this, Arrays.asList("public_profile", "email"))
+//        LoginManager.getInstance()
+//            .registerCallback(callbackManager, object : FacebookCallback<LoginResult> {
+//                override fun onSuccess(result: LoginResult?) {
+//                    firebaseAuthWithFacebook(result)
+//                }
+//
+//                override fun onCancel() {
+////                TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+//                    Log.i("INFORMATION FACEBOOK", "CANCEL")
+//                }
+//
+//                override fun onError(error: FacebookException?) {
+////                TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+//                    Log.i("INFORMATION FACEBOOK", "ERROR")
+//                }
+//
+//            })
+//    }
 
-                override fun onCancel() {
-//                TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-                    Log.i("INFORMATION FACEBOOK", "CANCEL")
-                }
+//    fun firebaseAuthWithFacebook(result: LoginResult?) {
+//        var credential = FacebookAuthProvider.getCredential(result!!.accessToken.token)
+//        FirebaseAuth.getInstance().signInWithCredential(credential).addOnCompleteListener { task ->
+//            if (task.isSuccessful) {
+//                println("Facebook Login Success")
+//                val user = auth.currentUser
+//                val message = " ${user?.displayName} + ${user?.email} + ${user?.photoUrl}"
+//                Log.i("INFORMATION FACEBOOK", message)
+//            }
+//        }
+//    }
 
-                override fun onError(error: FacebookException?) {
-//                TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-                    Log.i("INFORMATION FACEBOOK", "ERROR")
-                }
 
-            })
-    }
 
-    fun firebaseAuthWithFacebook(result: LoginResult?) {
-        var credential = FacebookAuthProvider.getCredential(result!!.accessToken.token)
-        FirebaseAuth.getInstance().signInWithCredential(credential).addOnCompleteListener { task ->
-            if (task.isSuccessful) {
-                println("Facebook Login Success")
-                val user = auth.currentUser
-                val message = " ${user?.displayName} + ${user?.email} + ${user?.photoUrl}"
-                Log.i("INFORMATION FACEBOOK", message)
-            }
-        }
-    }
-
-    private fun launchSignInFlow() {
-        // Give users the option to sign in / register with their email, Google account, or Facebook account.
-        // If users choose to register with their email,
-        // they will need to create a password as well.
-        val providers = arrayListOf(
-            AuthUI.IdpConfig.EmailBuilder().build(),
-            AuthUI.IdpConfig.GoogleBuilder().build(),
-            AuthUI.IdpConfig.FacebookBuilder().build()
-
-            // This is where you can provide more ways for users to register and
-            // sign in.
-        )
-
-        // Create and launch sign-in intent.
-        // We listen to the response of this activity with the
-        // SIGN_IN_REQUEST_CODE
-        startActivityForResult(
-            AuthUI.getInstance()
-                .createSignInIntentBuilder()
-                .setAvailableProviders(providers)
-//                .setLogo(R.drawable.fui_ic_facebook_white_22dp)
-                .build(),
-            SIGN_IN_RESULT_CODE
-        )
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == SIGN_IN_RESULT_CODE) {
-            val response = IdpResponse.fromResultIntent(data)
-            if (resultCode == Activity.RESULT_OK) {
-
-                // For getting the user information
-                val user = FirebaseAuth.getInstance().currentUser
-                user?.let {
-                    name = user?.displayName.toString()
-                    val email = user.email
-                    photoUrl = user?.photoUrl.toString()
-                    userId = user?.uid
-
-                    Toast.makeText(fragmentContext, "Welcome $name", Toast.LENGTH_SHORT).show()
-
-                    Log.i("HomeFragment", name + email + photoUrl)
-
-                    // To get user's preference data using custom objects
-                    val docRef = db.collection("Users").document(userId!!)
-                    docRef.get().addOnSuccessListener { documentSnapshot ->
-                        this.user = documentSnapshot.toObject(User::class.java)!!
-                        Log.i("HomeFragment2", this.user?.toString())
-                    }
-                }
-
-                // User successfully signed in
-                Log.i(TAG, "Successfully signed in user ${FirebaseAuth.getInstance().currentUser?.displayName}!")
-            } else {
-                // Sign in failed. If response is null the user canceled the
-                // sign-in flow using the back button. Otherwise check
-                // response.getError().getErrorCode() and handle the error.
-                Log.i(TAG, "Sign in unsuccessful ${response?.error?.errorCode}")
-            }
-        }
-    }
 
     /**
      * Inflates the overflow menu that contains filtering options.
@@ -416,7 +303,7 @@ class HomeFragment : Fragment() {
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
 
         when (item.itemId) {
-            R.id.preferences -> findNavController().navigate(HomeFragmentDirections.actionHomeFragmentToSettingsFragment(name, photoUrl, user))
+            R.id.preferences -> findNavController().navigate(HomeFragmentDirections.actionHomeFragmentToSettingsFragment(viewModel.name, viewModel.photo, viewModel.user))
         }
 
         return super.onOptionsItemSelected(item)
