+++ app/src/main/res/layout/fragment_marketing.xml
@@ -4,7 +4,9 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
-
+        <variable
+            name="viewModel"
+            type="com.lambda.mnemecards.marketing.MarketingViewModel" />
     </data>
 
     <ScrollView
